feat: [#36] Created index.ts files for default export and renamed file names
