+++ src/apis/getItem.ts
@@ -1,5 +1,5 @@
 import getData from '@/helpers/getData';
-import { ContentType, IItem } from '@/types/types';
+import { ContentType, IItem } from '@/types/shared';
 
 const formatUrl = (contentType: ContentType, id: string): string =>
   `${process.env.REACT_APP_MOVIES_API_BASE_URL}${contentType}/${id}?api_key=${process.env.REACT_APP_MOVIES_API_KEY}&append_to_response=videos`;
