+++ src/components/Navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 import React, { useContext, useState } from 'react';
 
 import { CONTENT_TYPE, SHOW_PLACEHOLDER, MOVIE_PLACEHOLDER } from '@/constants/constantValues';
-import { MoviesShowsDispatchContext, MoviesShowsContext } from '@/context/Context';
+import { MoviesShowsDispatchContext, MoviesShowsContext } from '@/context';
 
 import './Navbar.css';
 import { INavbar } from './types';
